+++ resources/404.php
@@ -1,2 +1,0 @@
-<h1>Unauthorized</h1>
-<p>You do not have sufficient permissions. Please <a href='index.html'>login</a> and try to access this page again.</p>