+++ resources/adminControls.php
@@ -1,4 +1,4 @@
-<?php 
+<?php
 // Open the session socket to begin testing session variables 
 session_start();
 
@@ -6,8 +6,9 @@ if( !isset($_SESSION['NAME']) ){
     header('HTTP/1.0 401 Unauthorized');
     include_once '401.php';
     exit();
-} 
+}
 else {
-    include_once 'copyCardTemplate.php';
+    include_once 'placeholder.php';
 }
 ?>
+