+++ resources/placeholder.php
@@ -1,0 +1,3 @@
+<?php
+echo "<h3>Uh oh. We're missing something here. Why don't we wait and let the dev see if he is going to implement something cool!</h3>";
+?>